# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  auth.LoginResponse:
    description: Login response
    properties:
      message:
        description: |-
          Success message
          @Example "Login successful"
        type: string
    type: object
  auth.LogoutResponse:
    description: Logout response
    properties:
      message:
        description: |-
          Success message
          @Example "Logout successful"
        type: string
    type: object
  auth.UserResponse:
    description: User information response
    properties:
      created_at:
        description: |-
          Time when the user was created
          @Example "2024-01-01T00:00:00Z"
        type: string
      last_login_at:
        description: |-
          Last time the user logged in
          @Example "2024-01-01T12:34:56Z"
        type: string
      username:
        description: |-
          Username of the user
          @Example "admin"
        type: string
    type: object
...
  handler.CreateOrganizationRequest:
    properties:
      description:
        type: string
      mspId:
        type: string
      name:
        type: string
      providerId:
        type: integer
    required:
    - mspId
    - name
    type: object
  handler.OrganizationResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      mspId:
        type: string
      providerId:
        type: integer
      providerName:
        type: string
      signCertificate:
        type: string
      signPublicKey:
        type: string
      tlsCertificate:
        type: string
      tlsPublicKey:
        type: string
      updatedAt:
        type: string
    type: object
  handler.UpdateOrganizationRequest:
    properties:
      description:
        type: string
    type: object
...
host: localhost:8100
info:
  contact:
    email: support@chainlaunch.com
    name: API Support
    url: http://www.chainlaunch.com/support
  description: ChainLaunch API provides services for managing blockchain networks
    and cryptographic keys
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ChainLaunch API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a session cookie
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Login user
      tags:
      - Authentication
  /auth/logout:
    post:
      description: Invalidates the current session and clears the session cookie
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/auth.LogoutResponse'
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Logout user
      tags:
      - Authentication
  /auth/me:
    get:
      description: Returns information about the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/auth.UserResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - CookieAuth: []
      summary: Get current user
      tags:
      - Authentication
...
  /organizations:
    get:
      consumes:
      - application/json
      description: Get a list of all Fabric organizations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.OrganizationResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all Fabric organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Create a new Fabric organization with the specified configuration
      parameters:
      - description: Organization creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new Fabric organization
      tags:
      - organizations
  /organizations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Fabric organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a Fabric organization
      tags:
      - organizations
    get:
      consumes:
      - application/json
      description: Get a Fabric organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OrganizationResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a Fabric organization
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Update an existing Fabric organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      - description: Organization update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a Fabric organization
      tags:
      - organizations
  /organizations/by-mspid/{mspid}:
    get:
      consumes:
      - application/json
      description: Get a Fabric organization by MSP ID
      parameters:
      - description: MSP ID
        in: path
        name: mspid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OrganizationResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a Fabric organization by MSP ID
      tags:
      - organizations
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  CookieAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"
...